refactor(mha): another half of mha refactor
